refactored neuron class structure. added input and bias neurons. added neuron layer. added tests for bias neuron

+++ src/com/joefrew/neuralnet/WeightedSynapse.java
@@ -10,7 +10,7 @@ public class WeightedSynapse extends Synapse {
 	
 	public WeightedSynapse() {}
 
-	public WeightedSynapse(int weight) {
+	public WeightedSynapse(double weight) {
 		this.weight = weight;
 	}
